// src/modules/attachments/gRPC/proto/storage.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.0
//   protoc               v3.12.4
// source: storage.proto

/* eslint-disable */

export const protobufPackage = 'storage';

export interface SaveFileRequest {
    userId: string;
    filename: string;
    fileContent: Uint8Array;
}

export interface SaveFileResponse {
    success: boolean;
    message: string;
    storedPath: string;
}

export interface GetFileRequest {
    userId: string;
    filename: string;
}

export interface GetFileResponse {
    success: boolean;
    message: string;
    fileContent: Uint8Array;
}

export interface DeleteFileRequest {
    userId: string;
    filename: string;
}

export interface DeleteFileResponse {
    success: boolean;
    message: string;
}